# Conflicts:
#	bkmonitor/api/apm_api/default.py
#	bkmonitor/apm/models/datasource.py
#	bkmonitor/apm/resources.py
#	bkmonitor/apm/task/tasks.py
#	bkmonitor/packages/apm_web/profile/resources.py
#	bkmonitor/packages/apm_web/profile/urls.py
#	bkmonitor/packages/apm_web/profile/views.py

+++ bkmonitor/webpack/src/monitor-pc/pages/alarm-dispatch/alarm-dispatch.tsx
@@ -61,6 +61,8 @@ export default class AlarmDispatch extends tsc<{}> {
   cacheRuleGroups = [];
   /* 搜索 */
   search = '';
+  /** 是否能查询，需要依赖 getAlarmGroupList 和 getAlarmDispatchGroupData两个函数都请求完成 */
+  isSearch = false;
   /** 新建规则组弹窗 */
   visible = false;
 
@@ -176,18 +178,21 @@ export default class AlarmDispatch extends tsc<{}> {
   }
 
   created() {
+    this.getRouteParams();
     this.getAlarmDispatchGroupData();
     this.getAlarmGroupList();
     this.getKVOptionsData();
     this.getProcessPackage();
     this.$store.commit('app/SET_NAV_ROUTE_LIST', [{ name: this.$t('route-告警分派'), id: '' }]);
-    this.getRouteParams();
   }
 
   getRouteParams() {
-    const { groupName } = this.$route.params;
+    const { groupName, group_id: groupId } = this.$route.query;
     if (groupName) {
-      this.search = groupName;
+      this.search = groupName as string;
+    }
+    if (groupId) {
+      this.search = groupId as string;
     }
   }
 
@@ -199,6 +204,12 @@ export default class AlarmDispatch extends tsc<{}> {
       name: item.name,
       receiver: item.users?.map(rec => rec.display_name) || []
     }));
+
+    /** 能否查询需要依赖 getAlarmGroupList 和 getAlarmDispatchGroupData两个函数都请求完成 */
+    if (this.isSearch && this.search) {
+      this.handleSearch();
+    }
+    this.isSearch = true;
   }
 
   /**
@@ -217,18 +228,19 @@ export default class AlarmDispatch extends tsc<{}> {
             name: item.name,
             priority: item.priority,
             isExpan: true,
-            ruleData: []
+            ruleData: [],
+            editAllowed: !!item?.edit_allowed
           })
       ) || [];
     this.loading = false;
-    this.getAlarmAssignGroupsRules(list.map(item => item.id));
+    this.getAlarmAssignGroupsRules(list?.map(item => item.id));
   }
 
   /**
    *  获取规则集
    * @param ids 规则组
    */
-  async getAlarmAssignGroupsRules(ids: number[]) {
+  async getAlarmAssignGroupsRules(ids: number[] = []) {
     this.groupLoading = true;
     const list = await listAssignRule().catch(() => (this.groupLoading = false));
     const groupData = ids.reduce((result, item) => {
@@ -242,7 +254,10 @@ export default class AlarmDispatch extends tsc<{}> {
     });
     this.cacheRuleGroups = this.ruleGroups;
     this.loading = false;
-    this.search && this.handleSearch();
+    if (this.isSearch && this.search) {
+      this.handleSearch();
+    }
+    this.isSearch = true;
   }
 
   // 获取流程套餐
@@ -478,7 +493,8 @@ export default class AlarmDispatch extends tsc<{}> {
             item.tag.map(key => key.split(':')[0]).includes(value) ||
             item.tag.map(value => value.split(':')[1]).includes(value) ||
             reg.test(item.user_groups.map(g => g.name)) ||
-            item.user_groups.map(g => String(g.id)).includes(value)
+            String(item.id) === value
+          // item.user_groups.map(g => String(g.id)).includes(value)
         )
         .map(item => item.id);
       this.ruleGroups = this.cacheRuleGroups.filter(item => filterRuleGroupList.includes(item.id));
@@ -486,6 +502,12 @@ export default class AlarmDispatch extends tsc<{}> {
     } else {
       this.emptyText = window.i18n.tc('查无数据');
       this.ruleGroups = this.cacheRuleGroups;
+      const { groupName, groupId, ...arg } = this.$route.query;
+      this.$router.replace({
+        query: {
+          ...arg
+        }
+      });
     }
   }
 
@@ -591,16 +613,26 @@ export default class AlarmDispatch extends tsc<{}> {
                     {this.renderEditAttribute(item.name, item.ruleData.length, 'name', item.id)}
                     {this.renderEditAttribute(`${this.$t('优先级')}:`, item.priority, 'priority', item.id)}
                     <div
-                      class='edit-btn-wrap'
-                      onClick={() => this.handleToConfig(item.id)}
+                      class={['edit-btn-wrap', { 'edit-btn-disabled': !item.editAllowed }]}
+                      v-bk-tooltips={{
+                        placements: ['top'],
+                        content: this.$t('内置的分派规则组不允许修改'),
+                        disabled: item.editAllowed
+                      }}
+                      onClick={() => item.editAllowed && this.handleToConfig(item.id)}
                     >
                       <span class='icon-monitor icon-bianji'></span>
                       <span>{this.$t('配置规则')}</span>
                     </div>
                     <div
-                      class='del-btn-wrap'
+                      class={['del-btn-wrap', { 'del-btn-disabled': !item.editAllowed }]}
+                      v-bk-tooltips={{
+                        placements: ['top'],
+                        content: this.$t('内置的分派规则组不允许修改'),
+                        disabled: item.editAllowed
+                      }}
                       onClick={e => {
-                        this.handleDeleteGroup(e, item);
+                        item.editAllowed && this.handleDeleteGroup(e, item);
                       }}
                     >
                       <span class='icon-monitor icon-mc-delete-line'></span>
@@ -669,6 +701,7 @@ export default class AlarmDispatch extends tsc<{}> {
                               detailData={this.detailData}
                               processPackage={this.processPackage}
                               actions={row.actions}
+                              userType={row.user_type}
                             />
                           )
                         }}
